+++ CHANGES.rst
@@ -7,6 +7,9 @@ General
 New Features
 ^^^^^^^^^^^^
 
+- Inclusion of galaxy count model, to be used in the generation of perturbation
+  AUF components. [#41]
+
 - Creation of initial Galactic proper motion model, for inclusion within the
   cross-match framework in a future release. [#39]
 
@@ -108,6 +111,8 @@ API Changes
 
 Other Changes
 ^^^^^^^^^^^^^
+
+- Added ``skypy`` and ``speclite`` as dependencies. [#41]
 
 - Improved github actions matrix testing coverage. [#40]